+++ esp-hal-common/build.rs
@@ -156,6 +156,10 @@ fn main() -> Result<(), Box<dyn Error>> {
     let atomic_emulation_enabled = cfg!(feature = "atomic-emulation");
     let portable_atomic_enabled = cfg!(feature = "portable-atomic");
 
+    if !atomic_emulation_enabled && detect_atomic_extension("a") {
+        panic!("Atomic emulation flags detected in `.cargo/config.toml` but `atomic-emulation` feature is not enabled!");
+    }
+
     if atomic_emulation_enabled && portable_atomic_enabled {
         panic!("The `atomic-emulation` and `portable-atomic` features cannot be used together");
     }
@@ -204,12 +208,14 @@ fn main() -> Result<(), Box<dyn Error>> {
     println!("cargo:rustc-link-search={}", out.display());
 
     if cfg!(feature = "esp32") || cfg!(feature = "esp32s2") || cfg!(feature = "esp32s3") {
-        fs::copy("ld/xtensa/hal-defaults.x", out.join("hal-defaults.x")).unwrap();
+        fs::copy("ld/xtensa/hal-defaults.x", out.join("hal-defaults.x"))?;
+
         let (irtc, drtc) = if cfg!(feature = "esp32s3") {
             ("rtc_fast_seg", "rtc_fast_seg")
         } else {
             ("rtc_fast_iram_seg", "rtc_fast_dram_seg")
         };
+
         let alias = format!(
             r#"
             REGION_ALIAS("ROTEXT", irom_seg);
@@ -221,7 +227,8 @@ fn main() -> Result<(), Box<dyn Error>> {
         "#,
             irtc, drtc
         );
-        fs::write(out.join("alias.x"), alias).unwrap();
+
+        fs::write(out.join("alias.x"), alias)?;
     } else {
         fs::copy("ld/riscv/hal-defaults.x", out.join("hal-defaults.x"))?;
         fs::copy("ld/riscv/asserts.x", out.join("asserts.x"))?;
@@ -229,6 +236,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     }
 
     copy_dir_all("ld/sections", &out)?;
+    copy_dir_all(format!("ld/{device_name}"), &out)?;
 
     // Generate the eFuse table from the selected device's CSV file:
     gen_efuse_table(device_name, out)?;
@@ -313,4 +321,34 @@ fn gen_efuse_table(device_name: &str, out_dir: impl AsRef<Path>) -> Result<(), B
     }
 
     Ok(())
+}
+
+fn detect_atomic_extension(ext: &str) -> bool {
+    let rustflags = env::var_os("CARGO_ENCODED_RUSTFLAGS")
+        .unwrap()
+        .into_string()
+        .unwrap();
+
+    // Users can pass -Ctarget-feature to the compiler multiple times, so we have to
+    // handle that
+    let target_flags = rustflags
+        .split(0x1f as char)
+        .filter(|s| s.starts_with("target-feature="))
+        .map(|s| s.strip_prefix("target-feature="))
+        .flatten();
+    for tf in target_flags {
+        let tf = tf
+            .split(",")
+            .map(|s| s.trim())
+            .filter(|s| s.starts_with('+'))
+            .map(|s| s.strip_prefix('+'))
+            .flatten();
+        for tf in tf {
+            if tf == ext {
+                return true;
+            }
+        }
+    }
+
+    false
 }