+++ esp32c3-hal/examples/embassy_i2s_sound.rs
@@ -54,7 +54,7 @@ const SINE: [i16; 64] = [
 ];
 
 #[main]
-async fn main(_spawner: Spawner) -> ! {
+async fn main(_spawner: Spawner) {
     #[cfg(feature = "log")]
     esp_println::logger::init_logger_from_env();
     println!("Init!");