+++ esp32h2-hal/examples/embassy_spi.rs
@@ -36,7 +36,7 @@ use esp32h2_hal::{
 use esp_backtrace as _;
 
 #[main]
-async fn main(_spawner: Spawner) -> ! {
+async fn main(_spawner: Spawner) {
     esp_println::println!("Init!");
     let peripherals = Peripherals::take();
     let system = peripherals.SYSTEM.split();